+++ src/main/aerospike.cc
@@ -78,8 +78,7 @@ NAN_METHOD(client)
 {
 	Nan::HandleScope();
 	Local<Object> config = info[0].As<Object>();
-	Local<Function> eventCb = info[1].As<Function>();
-	info.GetReturnValue().Set(AerospikeClient::NewInstance(config, eventCb));
+	info.GetReturnValue().Set(AerospikeClient::NewInstance(config));
 }
 
 /**