+++ History.md
@@ -3,6 +3,7 @@ HEAD
 
 * **Bug Fixes**
   * Support all data types for write operation in Client#operate [#235](https://github.com/aerospike/aerospike-client-nodejs/issues/235)
+  * Only setup cluster events callback on Client#connect [#237](https://github.com/aerospike/aerospike-client-nodejs/issues/237)
 
 v3.1.0 / 2017-12-18
 ===================